pyproject.toml
[tool.poetry]
name = "flask-google-oauth-app"
version = "0.1.0"
description = "A Flask app with Google OAuth and local authentication."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0"
Flask-Login = "^0.6"
Flask-SQLAlchemy = "^3.0"
python-dotenv = "^1.0"
google-auth = "^2.0"
google-auth-oauthlib = "^1.0"
google-api-python-client = "^2.0"
Werkzeug = "^2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 